requirements.txt
streamlit
pandas
sqlite3
plotly
python-dotenv